Implements dom purification in reusable module - apply to RichText and deprecated RandomStyle field
