+++ app/src/gui/fields/RichText.tsx
@@ -17,187 +17,18 @@
 /**
  * RichTextField Component
  * 
- * A secure React component for rendering markdown content with HTML support.
- * This component implements multiple layers of security to safely render user-generated
- * content while preventing XSS attacks and other security vulnerabilities.
- * 
- * Security Features:
- * 1. HTML Sanitization using DOMPurify
- * 2. Strict tag and attribute whitelisting
- * 3. Protocol validation for links
- * 4. Domain validation for images
- * 5. Forced security attributes for external links
- * 
- * @security This component uses dangerouslySetInnerHTML, but implements multiple 
- * safeguards to prevent XSS attacks. All HTML is sanitized before rendering.
+ * Safely injects markdown content directly into HTML - see DomPurifier.ts
  */
 
+import {contentToSanitizedHtml} from '../../utils/DomPurifier';
 import React from 'react';
-import MarkdownIt from 'markdown-it';
-import DOMPurify from 'dompurify';
-import type {Config} from 'dompurify';
-
-// Configure markdown-it with minimal HTML support
-// html: true is required for img tags to work, but all HTML is sanitized before rendering
-// typographer: true enables smart quotes and other typographic features
-const md = new MarkdownIt({
-  html: true,
-  typographer: true,
-});
-
-/**
- * Validates URLs to ensure they use safe protocols.
- * This prevents javascript: URLs and other potentially malicious protocols.
- * Allows base 64 images
- *
- * @param url - The URL to validate
- * @returns boolean - Whether the URL uses an allowed protocol
- */
-md.validateLink = (url: string) => {
-  const protocols = ['http:', 'https:', 'mailto:', 'tel:', 'data:'];
-  try {
-    const urlObj = new URL(url);
-    return protocols.includes(urlObj.protocol);
-  } catch {
-    // Special handling for data URLs which might not parse as URLs
-    if (url.startsWith('data:image/')) {
-      return true;
-    }
-    return false;
-  }
-};
-
-/**
- * DOMPurify configuration
- * Implements a strict security policy for HTML sanitization
- */
-const purifyConfig: Config = {
-  // Explicitly whitelist allowed HTML tags
-  // Any tags not in this list will be stripped from the output
-  ALLOWED_TAGS: [
-    // Text formatting - Basic typography and code display
-    'p',
-    'br',
-    'b',
-    'i',
-    'em',
-    'strong',
-    'code',
-    'pre',
-
-    // Structure - Basic document structure elements
-    'div',
-    'span',
-    'h1',
-    'h2',
-    'h3',
-    'h4',
-    'h5',
-    'h6',
-
-    // Lists - Ordered and unordered lists
-    'ul',
-    'ol',
-    'li',
-
-    // Media - Only images are allowed
-    'img',
-
-    // Links - External links with forced security attributes
-    'a',
-  ],
-
-  // Explicitly whitelist allowed attributes
-  // Any attributes not in this list will be stripped
-  ALLOWED_ATTR: [
-    // General attributes - Basic styling and identification
-    'id',
-    'class',
-
-    // Media attributes - Required for images
-    'src',
-    'alt',
-    'title',
-
-    // Link attributes - Required for anchors
-    'href',
-    'target',
-    'rel',
-  ],
-
-  // Disable data attributes to prevent potential DOM-based XSS
-  ALLOW_DATA_ATTR: false,
-
-  // Enable DOM sanitization
-  SANITIZE_DOM: true,
-
-  // Don't treat input as a complete HTML document
-  WHOLE_DOCUMENT: false,
-};
-
-/**
- * Hook to create a configured instance of DOMPurify
- */
-const createSanitizer = () => {
-  const purify = DOMPurify(window);
-
-  purify.addHook('afterSanitizeAttributes', node => {
-    // Handle links
-    if (node.tagName === 'A' && node.hasAttribute('href')) {
-      node.setAttribute('rel', 'noopener noreferrer');
-      node.setAttribute('target', '_blank');
-    }
-
-    // Handle images
-    if (node.tagName === 'IMG' && node.hasAttribute('src')) {
-      const src = node.getAttribute('src') || '';
-
-      // Handle data URLs for images
-      if (src.startsWith('data:image/')) {
-        // Validate data URL format
-        const isValidDataUrl =
-          /^data:image\/(jpeg|jpg|png|gif|webp|svg\+xml);base64,([a-zA-Z0-9+/]+={0,2})$/.test(
-            src
-          );
-
-        if (!isValidDataUrl) {
-          node.removeAttribute('src');
-        }
-        // Data URL is valid, allow it to remain
-        return;
-      }
-
-      // Handle regular URLs
-      try {
-        const url = new URL(src);
-        const trustedDomains = ['yourdomain.com', 'cdn.yourdomain.com'];
-        if (!trustedDomains.some(domain => url.hostname.endsWith(domain))) {
-          node.removeAttribute('src');
-        }
-      } catch {
-        node.removeAttribute('src');
-      }
-    }
-  });
-
-  return purify;
-};
-
-// Get a configured DOMPurify instance
-const purify = createSanitizer();
-
 interface Props {
   /** The markdown content to be rendered. May include safe HTML tags. */
   content: string;
 }
 
 export const RichTextField: React.FC<Props> = ({content}) => {
-  // Process the content:
-  // 1. Convert markdown to HTML
-  // 2. Sanitize the HTML
-  // 3. Render the sanitized content
-  const renderedContent = md.render(content);
-  const sanitizedContent = purify.sanitize(renderedContent, purifyConfig);
-
-  return <div dangerouslySetInnerHTML={{__html: sanitizedContent}} />;
+  return (
+    <div dangerouslySetInnerHTML={{__html: contentToSanitizedHtml(content)}} />
+  );
 };
